Display the books, received from the slice, in a reusable component.
