+++ src/pages/Books/BooksPage.jsx
@@ -1,33 +1,5 @@
-import { useSelector } from 'react-redux';
-import Book from '../../Components/Book/Book';
-import Form from '../../Components/Form/Form';
+import BookList from '../../Components/BookList/BookList';
 
-const BooksPage = () => {
-  const books = useSelector((store) => store.books);
-
-  if (books.length === 0) {
-    return (
-      <>
-        <h3>Library is empty....</h3>
-        <Form />
-      </>
-    );
-  }
-  return (
-    <div
-      style={{
-        display: 'flex',
-        flexDirection: 'column',
-        gap: '20px',
-        padding: '50px',
-      }}
-    >
-      {books?.map((book) => (
-        <Book key={book.id} title={book.title} author={book.author} />
-      ))}
-      <Form />
-    </div>
-  );
-};
+const BooksPage = () => <BookList />;
 
 export default BooksPage;
